Properly restore export bit
Changes to fish 3.0 meant that doing

```
set -g LANG $LANG
```

would cause it to fall back to the C locale (i.e. no unicode).

However, fishtape_restore_globals unset the export bit for every
variable, which is wrong in general.

So we save the export state in the variable name and restore it later.